Added basic buddy list operation (single player only!) (NOT FOR PRODUCTION USE)
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@1449 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/PlayerBuddyList.cpp
@@ -16,8 +16,10 @@ along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
 #include "PlayerBuddyList.h"
+#include "BuddyListPacket.h"
 #include "Player.h"
 #include "MySQLM.h"
+#include <algorithm>
 
 PlayerBuddyList::PlayerBuddyList(Player *player) : player(player) {
 	mysqlpp::Query query = Database::getCharDB().query();
@@ -29,40 +31,72 @@ PlayerBuddyList::PlayerBuddyList(Player *player) : player(player) {
 	}
 }
 
-void PlayerBuddyList::add(int32_t charid) {
-	mysqlpp::Query query = Database::getCharDB().query();
-	query << "INSERT INTO buddylist (charid, buddy_charid) VALUES (" << player->getId() << ", " << charid << ")";
-	mysqlpp::SimpleResult res = query.execute();
-	
-	query << "SELECT buddylist.buddy_charid as buddy_charid, characters.name as name FROM buddylist INNER JOIN characters ON buddylist.buddy_charid = characters.id WHERE buddylist.id = " << res.insert_id();
-	mysqlpp::StoreQueryResult res2 = query.store();
+uint8_t PlayerBuddyList::add(const string &name) {
+	if (buddies.size() >= player->getBuddyListSize()) { // Buddy list full
+		return 0x0b;
+	}
 
-	add(res2[0]);
-}
-
-bool PlayerBuddyList::add(const string &name) {
 	mysqlpp::Query query = Database::getCharDB().query();
-	query << "SELECT id FROM characters WHERE name = " << mysqlpp::quote << name;
+	query << "SELECT c.id, u.gm, c.buddylist_size AS buddylist_limit,(SELECT COUNT(b.id) FROM buddylist b WHERE b.charid = c.id ) AS buddylist_size FROM characters c INNER JOIN users u ON c.userid = u.id WHERE c.name = " << mysqlpp::quote << name;
 	mysqlpp::StoreQueryResult res = query.store();
 
 	if (res.size() == 0) { // Name does not exist
-		return false;
+		return 0x0f;
 	}
 
-	add(res[0][0]);
-	return true;
+	if ((int32_t) res[0][1] > 0 && !player->isGM()) { // GM cannot be in buddy list unless the player is a GM
+		return 0x0e;
+	}
+
+	if ((uint32_t) res[0][3] >= (uint32_t) res[0][2]) { // Opposite-end buddy list full
+		return 0x0c;
+	}
+
+	int32_t charid = res[0][0];
+
+	if (buddies.find(charid) != buddies.end()) { // Already in buddy list
+		return 0x0d;
+	}
+
+	query << "INSERT INTO buddylist (charid, buddy_charid) VALUES (" << player->getId() << ", " << charid << ")";
+	mysqlpp::SimpleResult res2 = query.execute();
+	
+	query << "SELECT buddylist.buddy_charid as buddy_charid, characters.name as name FROM buddylist INNER JOIN characters ON buddylist.buddy_charid = characters.id WHERE buddylist.id = " << res2.insert_id();
+	res = query.store();
+
+	add(res[0]);
+
+	BuddyListPacket::update(player, BuddyListPacket::add);
+
+	return 0;
 }
 
 void PlayerBuddyList::remove(int32_t charid) {
+	if (buddies.find(charid) == buddies.end()) {
+		// Hacking
+		return;
+	}
+
 	mysqlpp::Query query = Database::getCharDB().query();
 	query << "DELETE FROM buddylist WHERE charid = " << player->getId() << " AND buddy_charid = " << charid;
 	query.exec();
 
 	buddies.erase(charid);
+	buddies_order.erase(std::remove(buddies_order.begin(), buddies_order.end(), charid));
+
+	BuddyListPacket::update(player, BuddyListPacket::remove);
 }
 
 inline
 void PlayerBuddyList::add(const mysqlpp::Row &row) {
 	int32_t charid = (int32_t) row["buddy_charid"];
-	row["name"].to_string(buddies[charid]); // Assigns string to buddies[charid] i.e. buddies[charid] = name
+
+	BuddyPtr buddy(new Buddy);
+	buddy->charid = charid;
+	row["name"].to_string(buddy->name);
+	buddy->oppositeStatus = OppositeStatus::unregistered; // FIXME
+	buddy->channel = -1; // FIXME
+
+	buddies[charid] = buddy;
+	buddies_order.push_back(charid);
 }
